refactor(@xen-orchestra/backups): move Backup , VmBackup,  writers and specific method to a private folder
