+++ @xen-orchestra/backups/_backupJob/VmBackup/VmBackup.js
@@ -16,13 +16,13 @@ const { pipeline } = require('node:stream')
 
 const { IncrementalBackupWriter } = require('./writers/IncrementalBackupWriter.js')
 const { IncrementalReplicationWriter } = require('./writers/IncrementalReplicationWriter.js')
-const { exportDeltaVm } = require('./_deltaVm.js')
-const { forkStreamUnpipe } = require('./_forkStreamUnpipe.js')
+const { exportDeltaVm } = require('../../_deltaVm.js')
+const { forkStreamUnpipe } = require('../forkStreamUnpipe.js')
 const { FullBackupWriter } = require('./writers/FullBackupWriter.js')
 const { FullReplicationWriter } = require('./writers/FullReplicationWriter.js')
-const { getOldEntries } = require('./_getOldEntries.js')
-const { Task } = require('./Task.js')
-const { watchStreamSize } = require('./_watchStreamSize.js')
+const { getOldEntries } = require('../../_getOldEntries.js')
+const { Task } = require('../../Task.js')
+const { watchStreamSize } = require('../../_watchStreamSize.js')
 
 const { debug, warn } = createLogger('xo:backups:VmBackup')
