+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/FullBackupWriter.js
@@ -1,8 +1,8 @@
 'use strict'
 
-const { formatFilenameDate } = require('../_filenameDate.js')
-const { getOldEntries } = require('../_getOldEntries.js')
-const { Task } = require('../Task.js')
+const { formatFilenameDate } = require('../../../_filenameDate.js')
+const { getOldEntries } = require('../../../_getOldEntries.js')
+const { Task } = require('../../../Task.js')
 
 const { MixinBackupWriter } = require('./_MixinBackupWriter.js')
 const { AbstractFullWriter } = require('./_AbstractFullWriter.js')
