+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/FullReplicationWriter.js
@@ -4,9 +4,9 @@ const ignoreErrors = require('promise-toolbox/ignoreErrors')
 const { asyncMap, asyncMapSettled } = require('@xen-orchestra/async-map')
 const { formatDateTime } = require('@xen-orchestra/xapi')
 
-const { formatFilenameDate } = require('../_filenameDate.js')
-const { getOldEntries } = require('../_getOldEntries.js')
-const { Task } = require('../Task.js')
+const { formatFilenameDate } = require('../../../_filenameDate.js')
+const { getOldEntries } = require('../../../_getOldEntries.js')
+const { Task } = require('../../../Task.js')
 
 const { AbstractFullWriter } = require('./_AbstractFullWriter.js')
 const { MixinReplicationWriter } = require('./_MixinReplicationWriter.js')
