+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/IncrementalReplicationWriter.js
@@ -4,10 +4,10 @@ const { asyncMap, asyncMapSettled } = require('@xen-orchestra/async-map')
 const ignoreErrors = require('promise-toolbox/ignoreErrors')
 const { formatDateTime } = require('@xen-orchestra/xapi')
 
-const { formatFilenameDate } = require('../_filenameDate.js')
-const { getOldEntries } = require('../_getOldEntries.js')
-const { importDeltaVm, TAG_COPY_SRC } = require('../_deltaVm.js')
-const { Task } = require('../Task.js')
+const { formatFilenameDate } = require('../../../_filenameDate.js')
+const { getOldEntries } = require('../../../_getOldEntries.js')
+const { importDeltaVm, TAG_COPY_SRC } = require('../../../_deltaVm.js')
+const { Task } = require('../../../Task.js')
 
 const { AbstractIncrementalWriter } = require('./_AbstractIncrementalWriter.js')
 const { MixinReplicationWriter } = require('./_MixinReplicationWriter.js')
