+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/_AbstractFullWriter.js
