+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/_AbstractIncrementalWriter.js
