+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/_AbstractWriter.js
