+++ @xen-orchestra/backups/_backupJob/getAdaptersByRemote.js
